# docs/api-documentation-openapi.json
{
    "openapi": "3.0.1",
...
    "info": {
        "title": "mimoto",
        "description": "mimoto",
        "license": {
            "name": "Mosip",
            "url": "https://docs.mosip.io/platform/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.collab.mosip.net/v1/mimoto",
            "description": "For Swagger"
        }
    ],
    "paths": {
        "/req/otp": {
            "post": {
                "operationId": "otpRequest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppOTPRequestDTO"
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "individualId": "1289358704",
                                        "individualIdType": "UIN",
                                        "otpChannel": [
                                            "EMAIL",
                                            "PHONE"
                                        ],
                                        "transactionID": "1234567890"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "version": {
                                            "type": "string"
                                        },
                                        "str": {
                                            "type": "null"
                                        },
                                        "responsetime": {
                                            "type": "null"
                                        },
                                        "metadata": {
                                            "type": "null"
                                        },
                                        "response": {
                                            "type": "object",
                                            "properties": {
                                                "maskedMobile": {
                                                    "type": "string"
                                                },
                                                "maskedEmail": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "x-examples": {
                                        "Example 1": {
                                            "id": "mosip.identity.otp.internal",
                                            "version": "1.0",
                                            "str": null,
                                            "responsetime": null,
                                            "metadata": null,
                                            "response": {
                                                "maskedMobile": "XXXXXX6677",
                                                "maskedEmail": "XXuXXrXX5XX9@gmail.com"
                                            },
                                            "errors": null
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": {
                                            "id": "mosip.identity.otp.internal",
                                            "version": "1.0",
                                            "str": null,
                                            "responsetime": null,
                                            "metadata": null,
                                            "response": {
                                                "maskedMobile": "XXXXXX6677",
                                                "maskedEmail": "XXuXXrXX5XX9@gmail.com"
                                            },
                                            "errors": null
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Request new otp",
                "tags": [
                    "Credential Download - OTP Flow"
                ],
                "summary": "Generate Credential - Request Otp"
            }
        },
        "/credentialshare/request": {
            "post": {
                "operationId": "request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "individualId": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    },
                                    "transactionID": {
                                        "type": "string"
                                    },
                                    "user": {
                                        "type": "string"
                                    }
                                },
                                "x-examples": {
                                    "Example 1": {
                                        "individualId": "2057135194",
                                        "otp": "111111",
                                        "transactionID": "1234567890",
                                        "user": "mono"
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "individualId": "2057135194",
                                        "otp": "111111",
                                        "transactionID": "1234567890",
                                        "user": "mono"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "version": {
                                            "type": "string"
                                        },
                                        "str": {
                                            "type": "null"
                                        },
                                        "responsetime": {
                                            "type": "string"
                                        },
                                        "metadata": {
                                            "type": "null"
                                        },
                                        "response": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "requestId": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "errors": {
                                            "type": "null"
                                        }
                                    },
                                    "x-examples": {
                                        "Example 1": {
                                            "id": "mosip.resident.vid",
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:08:14.846Z",
                                            "metadata": null,
                                            "response": {
                                                "id": "2057135194",
                                                "requestId": "2bc3bb7f-6156-46d1-ae03-bf7b76e0c257"
                                            },
                                            "errors": null
                                        }
                                    }
                                },
                                "examples": {
                                    "Success response": {
                                        "value": {
                                            "id": "mosip.resident.vid",
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:08:14.846Z",
                                            "metadata": null,
                                            "response": {
                                                "id": "2057135194",
                                                "requestId": "2bc3bb7f-6156-46d1-ae03-bf7b76e0c257"
                                            },
                                            "errors": null
                                        }
                                    },
                                    "Failure response": {
                                        "value": {
                                            "id": null,
                                            "version": "v1",
                                            "str": null,
                                            "responsetime": "2022-10-31T05:07:34.789Z",
                                            "metadata": null,
                                            "response": null,
                                            "errors": [
                                                {
                                                    "errorCode": "RES-SER-422",
                                                    "errorMessage": "Input transactionID does not match transactionID of OTP Request"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "Credential request",
                "tags": [
                    "Credential Download - OTP Flow"
                ],
                "summary": "Generate Credential -  request new credential"
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AppOTPRequestDTO": {
                "type": "object",
                "properties": {
                    "individualId": {
                        "type": "string"
                    },
                    "individualIdType": {
                        "type": "string"
                    },
                    "otpChannel": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "transactionID": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
...
}